codebase/agent/requirements.txt
flask
google-cloud-bigquery
google-cloud-aiplatform
vertexai
pytz
gunicorn
